+++ cisco_sdwan/base/catalog.py
@@ -62,6 +62,7 @@ _tag_dependency_list = [
     'policy_customapp',
     'policy_definition',
     'policy_list',
+    'tag_rule'
 ]